requirements.txt
streamlit>=1.32.0
opencv-python>=4.8.0.76
Pillow>=10.2.0
numpy>=1.24.0
matplotlib>=3.8.0      # For optional plotting or visual debugging
imutils>=0.5.4
pyvirtualcam>=0.5.0
